Allow the passing of server credentials as string instead of file path (#3)
Previously, server credentials for Google services had to be provided as a file path, which contains unencrypted security credentials. This commit introduces the ability to instantiate the ServiceCredentials class using the parseFromJSON method, allowing JSON strings to be passed directly as a parameter. This enables credentials to be stored  in environment variables instead of files.